requirements.txt
google-auth-oauthlib==1.0.0
google-auth==2.23.0
google-api-python-client==2.100.0
langdetect==1.0.9
deepl==1.16.1
openai==1.3.0
websockets==12.0
supabase==2.0.0